# swagger.yaml
openapi: 3.0.0
info:
  title: gas-sensor
  contact: {}
  version: "1.0"
servers:
  - url: http://ec2-13-50-241-92.eu-north-1.compute.amazonaws.com:3000/api/v1
    variables: {}
paths:
  /create-data:
    post:
      tags:
        - Misc
      summary: create new data
      operationId: createsensordata
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createSensorData"
            example:
              deviceId: "3as330"
              temperature: 30
              smoke: 99
              co: 30
              humidity: 84
              lpg: 30
              latitude: 380
              longitude: 80
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /data/{deviceId}/{dataId}:
    put:
      tags:
        - Misc
      summary: update sensor data record
      operationId: updatedata
      parameters:
        - name: deviceId
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
        - name: dataId
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/updateSensorData"
            example:
              temperature: 30
              smoke: 99
              co: 30
              humidity: 84
              lpg: 30
              latitude: 380
              longitude: 80
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
    delete:
      tags:
        - Misc
      summary: Delete Sensor data
      operationId: Deletegasdata
      parameters:
        - name: deviceId
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
        - name: dataId
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /devices/{deviceId}:
    delete:
      tags:
        - Misc
      summary: delete sensor device
      operationId: deletedevice
      parameters:
        - name: deviceId
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /data/{deviceId}:
    get:
      tags:
        - Misc
      summary: Get Sensor Data
      operationId: GetGasData
      parameters:
        - name: deviceId
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
components:
  schemas:
    createSensorData:
      title: createSensorData
      required:
        - deviceId
      type: object
      properties:
        deviceId:
          type: string
        temperature:
          type: string
        humidity:
          type: string
        co:
          type: string
        lpg:
          type: string
        smoke:
          type: string
        latitude:
          type: string
        longitude:
          type: string
    updateSensorData:
      title: updateSensorData
      type: object
      properties:
        temperature:
          type: string
        humidity:
          type: string
        co:
          type: string
        lpg:
          type: string
        smoke:
          type: string
        latitude:
          type: string
        longitude:
          type: string
tags:
  - name: Misc
    description: ""